(PC-8256) Avoid firebase warning
Warning: The measurement ID in the local Firebase config does not match the measurement ID fetched from the server.
We use firebase 8.0.0 and above 7.20.0, measurementId is dynamically fetched, but this is to avoid the warning and
serves as a fallback if fetching the config fails.

Reference: https://firebase.google.com/docs/analytics/get-started?platform=web